# refactor(sumologicexporter): optimize fields stringification