Use byte slices instead of strings where possible and modify a "stringified field byte slice" in place to prevent unnecessary copies.

Additionally short circuit in `fields.equals()` when attribute maps have different length to prevent using `cmp` for comparison (which is relatively expensive).

```
$ go test -v -count 1 -run ^$ -bench BenchmarkFields -benchmem -count 1 . > old.txt
$ gco -
Previous HEAD position was 86d1d22 tests(sumologicexporter): add fields stringification benchmark
Switched to branch 'optimize-fields'
$ go test -v -count 1 -run ^$ -bench BenchmarkFields -benchmem -count 1 . > new.txt
$ benchcmp old.txt new.txt
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark              old ns/op     new ns/op     delta
BenchmarkFields-16     997           638           -35.97%

benchmark              old allocs     new allocs     delta
BenchmarkFields-16     17             11             -35.29%

benchmark              old bytes     new bytes     delta
BenchmarkFields-16     464           368           -20.69%
```